refactor: enttity 리팩토링
