+++ src/main/java/com/connectCo/domain/Member/controller/AuthController.java
@@ -1,0 +1,4 @@
+package com.connectCo.domain.Member.controller;
+
+public class AuthController {
+}
