+++ src/main/java/com/connectCo/domain/Member/entity/LoginType.java
@@ -1,4 +1,4 @@
-package com.connectCo.domain.enums;
+package com.connectCo.domain.Member.entity;
 
 public enum LoginType {
     KAKAO, GOOGLE, NAVER
