+++ src/main/java/com/connectCo/domain/Member/entity/Member.java
@@ -1,0 +1,43 @@
+package com.connectCo.domain.Member.entity;
+
+import com.connectCo.global.utils.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.GenericGenerator;
+import org.hibernate.annotations.Where;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+@Getter
+@Entity
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Where(clause = "deleted_at is null")
+public class Member extends BaseEntity
+{
+    @Id
+    @GeneratedValue(generator = "uuid2")
+    @GenericGenerator(name = "uuid2")
+    @Column(columnDefinition = "BINARY(16)")
+    private UUID id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable =false)
+    private String phoneNumber;
+
+    @Enumerated(EnumType.STRING)
+    private LoginType loginType;
+
+    @Column(nullable = false)
+    private String refreshToken;
+
+    @JoinColumn
+    @Enumerated(EnumType.STRING)
+    @ElementCollection(fetch = FetchType.EAGER)
+    private List<Role> role = new ArrayList<>(List.of(Role.USER));
+}
