+++ src/main/java/com/connectCo/domain/Member/service/AuthService.java
@@ -1,0 +1,4 @@
+package com.connectCo.domain.Member.service;
+
+public interface AuthService {
+}
