+++ src/main/java/com/connectCo/domain/Member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.connectCo.domain.Member.service;
+
+public class MemberService {
+}
