--- src/main/java/com/connectCo/domain/enums/Category.java
@@ -1,5 +1,0 @@
-package com.connectCo.domain.enums;
-
-public enum Category {
-     족발,치킨,보쌈
-}
