--- src/main/java/com/connectCo/domain/enums/CouponType.java
@@ -1,5 +1,0 @@
-package com.connectCo.domain.enums;
-
-public enum CouponType {
-    VALID_COUNT, VALID_PERIOD, VALID_DATE
-}
