+++ src/main/java/com/connectCo/domain/group/entity/GroupType.java
@@ -1,4 +1,4 @@
-package com.connectCo.domain.enums;
+package com.connectCo.domain.group.entity;
 
 public enum GroupType {
     GROUP, INDIVIDUAL
