+++ src/main/java/com/connectCo/domain/image/entity/Image.java
@@ -1,13 +1,11 @@
-package com.connectCo.domain.image;
+package com.connectCo.domain.image.entity;
 
-import com.connectCo.domain.baseentity.BaseEntity;
+import com.connectCo.global.utils.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.Where;
 
@@ -15,21 +13,17 @@ import java.util.UUID;
 
 @Getter
 @Entity
-@NoArgsConstructor
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
 @Where(clause = "deleted_at is null")
 public class Image extends BaseEntity {
     @Id
     @GeneratedValue(generator = "uuid2")
-    @GenericGenerator(name = "uuid2", strategy = "uuid2")
-    @Column(name = "image_id", columnDefinition = "BINARY(16)")
+    @GenericGenerator(name = "uuid2")
+    @Column(columnDefinition = "BINARY(16)")
     private UUID id;
 
     @Column(nullable = false)
     private String url;
-
-    @Builder
-    public Image(UUID id, String url) {
-        this.id = id;
-        this.url = url;
-    }
 }
