+++ src/main/java/com/connectCo/domain/store/entity/Store.java
@@ -1,0 +1,40 @@
+package com.connectCo.domain.store.entity;
+
+import com.connectCo.global.utils.BaseEntity;
+import com.connectCo.domain.Member.entity.Member;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.GenericGenerator;
+import org.hibernate.annotations.Where;
+
+import java.util.UUID;
+
+@Getter
+@Entity
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Where(clause = "deleted_at is null")
+public class Store extends BaseEntity {
+    @Id
+    @GeneratedValue(generator = "uuid2")
+    @GenericGenerator(name = "uuid2")
+    @Column(columnDefinition = "BINARY(16)")
+    private UUID id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn
+    private Member member;
+
+    @Column(nullable = false)
+    private String storeName;
+
+    private String storeDescription;
+
+    @Column(nullable = false)
+    private String storeNumber;
+
+    @Column(nullable = false)
+    private String address;
+
+}
